requirements.txt
pyspark==3.5.0
confluent-kafka==2.3.0
fastavro==1.9.0
dbt-bigquery==1.7.0
pandas>=2.0.0
python-dotenv==1.0.0
requests==2.31.0
Faker==20.1.0
sseclient-py==1.8.0
pymongo==4.6.0
streamlit==1.28.0
plotly==5.18.0
google-cloud-bigquery==3.13.0
